refactor(components): rename *Page.astro to *Content.astro
BlogPage, CvPage, KatasPage, SelfPage, ImsPage are page bodies
(orchestrators), not pages. The rename makes the role of each file
honest without touching the public URL structure.

Also updates test imports that referenced atoms/buttonStyles and
atoms/glassCardStyles (now shared/).
