+++ src/components/blog/BlogContent.astro
