+++ src/components/case-study/ImsContent.astro
