+++ src/components/case-study/SelfContent.astro
