+++ src/components/katas/KatasContent.astro
