+++ src/pages/blog/index.astro
@@ -1,4 +1,4 @@
 ---
-import BlogPage from '../../components/blog/BlogPage.astro';
+import BlogContent from '../../components/blog/BlogContent.astro';
 ---
-<BlogPage lang="es" />
+<BlogContent lang="es" />
