+++ src/pages/en/cv/index.astro
@@ -1,4 +1,4 @@
 ---
-import CvPage from '../../../components/cv/CvPage.astro';
+import CvContent from '../../../components/cv/CvContent.astro';
 ---
-<CvPage lang="en" />
+<CvContent lang="en" />
