+++ src/pages/en/katas/index.astro
@@ -1,6 +1,6 @@
 ---
 import Layout from '../../../layouts/Layout.astro';
-import KatasPage from '../../../components/katas/KatasPage.astro';
+import KatasContent from '../../../components/katas/KatasContent.astro';
 import { buildKatasCollectionSchema } from '../../../lib/schema-org';
 import { t } from '../../../i18n/utils';
 
@@ -13,5 +13,5 @@ const schema = buildKatasCollectionSchema(lang, canonicalUrl, metaTitle, metaDes
 
 <Layout title={metaTitle} description={metaDescription} lang={lang} ogImage="/og/og-katas.png" variant="dark-radial">
   <script slot="head" is:inline type="application/ld+json" set:html={JSON.stringify(schema)} />
-  <KatasPage lang={lang} />
+  <KatasContent lang={lang} />
 </Layout>
