+++ src/pages/en/portfolio/aitorevi-dev.astro
@@ -1,4 +1,4 @@
 ---
-import SelfPage from '../../../components/case-study/SelfPage.astro';
+import SelfContent from '../../../components/case-study/SelfContent.astro';
 ---
-<SelfPage lang="en" />
+<SelfContent lang="en" />
