+++ src/pages/portfolio/aitorevi-dev.astro
@@ -1,4 +1,4 @@
 ---
-import SelfPage from '../../components/case-study/SelfPage.astro';
+import SelfContent from '../../components/case-study/SelfContent.astro';
 ---
-<SelfPage lang="es" />
+<SelfContent lang="es" />
