+++ src/pages/portfolio/ims.astro
@@ -1,5 +1,5 @@
 ---
-import ImsPage from '../../components/case-study/ImsPage.astro';
+import ImsContent from '../../components/case-study/ImsContent.astro';
 ---
 
-<ImsPage />
+<ImsContent />
